move servers to subdir
